+++ e2etest/base/datastream.py
@@ -44,7 +44,9 @@ def test_datastream_from_device_to_server(device: Device, test_cfg: TestCfg):
         flush=True,
     )
     for key, value in test_cfg.mock_data.items():
-        device.send(test_cfg.interface_device_data, "/" + key, value, datetime.now(tz=timezone.utc))
+        device.send_individual(
+            test_cfg.interface_device_data, "/" + key, value, datetime.now(tz=timezone.utc)
+        )
         time.sleep(0.005)
 
     time.sleep(1)